Move CXCppInterOp.h to CppInterOp folder
