+++ unittests/CppInterOp/TypeReflectionTest.cpp
@@ -7,7 +7,7 @@
 #include "clang/Frontend/CompilerInstance.h"
 #include "clang/Sema/Sema.h"
 
-#include "clang-c/CXCppInterOp.h"
+#include "CppInterOp/CXCppInterOp.h"
 
 #include "gtest/gtest.h"
